Fixed Major Issue with Null Pointer Exception While Setting Ships
Turns out that changing an array while it is in use in Java is BAD PRACTICE and should be AVOIDED. Thanks to Christopher Nash and about 4 hours of debugging, the game is working properly and I can continue implementing the GUI. Eclipse also decided it hated how the whitespace looked on this file and had to change it. I'm sure it won't be the last time it does that.